Fix issue #18
The numbers originally lined up with the non-adjustable hotkeys.  Since hotkeys have long since been user configurable, the numbers are just there now to provide some positional index in the menu, so going over 9 is okay.

Since 1-9 do still happen to line up with the default Alt+# hotkeys, it is still kind of backwards compatible with the original purpose for a typical configuration with only relatively few presets.