Remove unversioned storage handler
It has been long enough that nobody has unversioned storage anymore.  And besides, this was actually kinda buggy since it prevented the intended code path for initializing storage.  Luckily the "wrong" code path just so happened to lead to the same end result...for now.  This was definitely a land mine that could have blow up in the future!